chore: apply clippy auto-fixes for idiomatic Rust
Apply `cargo clippy --fix` suggestions: use `is_multiple_of()`,
`is_empty()`, `is_some_and()`, remove redundant closures, unnecessary
references, unused imports, and useless type conversions.

+++ src/token_account_decoder.rs
@@ -292,7 +292,7 @@ fn parse_mint_extensions(
                 json!({ "delegate": pod_option_pubkey_to_string(&e.delegate) })
             }),
             ExtensionType::DefaultAccountState => ext_json!(state, "DefaultAccountState", DefaultAccountState, |e| {
-                let state_str = match u8::from(e.state) {
+                let state_str = match e.state {
                     0 => "Uninitialized", 1 => "Initialized", 2 => "Frozen", _ => "Unknown"
                 };
                 json!({ "state": state_str })
